feat: added filtering for eventsources
